requirements.txt
openai>=1.75.0
python-dotenv>=1.0.1